+++ CHANGES.md
@@ -6,11 +6,14 @@
 - Build:
   - Saves intermediate binary
 - Provision
+- Change `SpartaOptions` to `ExtendedOptions` to comply with golint
+- Change `cloudformation.CloudFormationResourceName` to `cloudformation.ResourceName` to comply with golint
 
   - Parameters for bucket, code, archive so that out of band changes allowed
 
 - Removed `HandleAWSLambda`
 - Add `step.APIGatewayTaskState` for to support [calling API Gateway](https://docs.aws.amazon.com/step-functions/latest/dg/connect-api-gateway.html)
+- Eliminate WorkflowHooks singletons
 - TODO:
   - Workflow hooks return (context, error) for more go idiomatic